Reconstruction for blackbox_1 in EqualitySaturation.cpp
This is not currently working. Sometimes, after reconstructing the blackboxed op, we get segfaults when constructing ops that reference the blakcboxed op. In other cases, the reconstruction of the blackboxed op itself leads to an error on the rust side, namely `get_shape` in `optimize.rs` complaining that the Id doesn't exist. This second error is quite bizarre, my working hypothesis is that it is caused by `add_or_get_val` not working as expected.